client/pyproject.toml
[project]
name = "sertn-avs"
version = "0.1.0"
description = "Sertn AVS Client"
authors = []
requires-python = ">=3.10,<3.13"
dependencies = [
  "typer>=0.9.0",
  "tqdm>=4.66.2",
  "pydantic>=2.6.3",
  "web3>=6.15.1",
  "pyyaml>=6.0.2",
  "eth-account>=0.13.5",
  "fastapi>=0.115.12",
  "uvicorn>=0.34.0",
  "packaging>=24.2",
  "torch==2.4.1",
  "ezkl==19.0.7",
  "onnxruntime>=1.21.1",
  "eth-abi>=5.2.0",
  "python-dotenv>=1.1.1",
  "gitpython>=3.1.45",
]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-asyncio>=0.23.5", "pytest-cov>=4.1.0"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["common", "avs_operator", "aggregator"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
pythonpath = ["src"]

[project.scripts]
sertn = "main:app"

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = ["httpx>=0.28.1", "pytest>=8.3.5", "pytest-mock>=3.14.1"]
